Revert "Footer"
