+++ public/index.html
@@ -13,10 +13,6 @@
     <link rel="icon" href="http://booklocal.in/wp-content/uploads/2017/07/cropped-171101-DEVCON3-BookLocal2-SQ-32x32.png" sizes="32x32">
     <link rel="icon" href="http://booklocal.in/wp-content/uploads/2017/07/cropped-171101-DEVCON3-BookLocal2-SQ-192x192.png" sizes="192x192">
     <link rel="apple-touch-icon-precomposed" href="http://booklocal.in/wp-content/uploads/2017/07/cropped-171101-DEVCON3-BookLocal2-SQ-180x180.png">
-    <script defer src="https://use.fontawesome.com/releases/v5.0.6/js/all.js"></script>
-
-
-
     <!--
       Notice the use of %PUBLIC_URL% in the tags above.
       It will be replaced with the URL of the `public` folder during the build.
