+++ src/components/Hero.jsx
@@ -13,6 +13,7 @@ class Hero extends Component {
     return (
       <div className={`hero-img ${image}`}>
         {/* <p className="hero-text">{this.props.heroText}</p> */}
+        <p>{heroText}</p>
       </div>
     );
   }
