+++ src/components/Reserve.jsx
@@ -19,7 +19,6 @@ if (typeof web3 !== 'undefined') {
 
 // stolen code zone ^^^
 
-// let RRAbi = require('../../ABIs/RoomRentingAbi.js');
 let RRAbi = require('../../abis/RoomRentingAbi.js');
 let RRAddress = require('../../contractAddress/rinkebyAddress.js');
 let RR = web3.eth.contract(RRAbi).at(RRAddress);
@@ -55,7 +54,7 @@ class Reserve extends Component{
     console.log(this.dateConverter(this.state.start));
     console.log(this.dateConverter(this.state.stop));
     console.log("Reserve fired!");
-    console.log("("+web3.toBigNumber(this.state.tokenId)+","+web3.toBigNumber(this.dateConverter(this.state.start))+","+web3.toBigNumber(this.dateConverter(this.state.stop))+","+web3.fromAscii(this.state.accessCode,32)+",{from: "+web3.eth.accounts[0]+", gas: 3000000})");
+    console.log("("+web3.toBigNumber(this.state.tokenId)+","+web3.toBigNumber(this.dateConverter(this.state.start))+","+web3.toBigNumber(this.dateConverter(this.state.stop))+","+web3.fromAscii(this.state.accessCode,32)+",{from: "+web3.eth.accounts[0]+", gas: 3000000}");
     reserve = RR.reserve(
       web3.toBigNumber(this.state.tokenId),
       web3.toBigNumber(this.dateConverter(this.state.start)),
@@ -84,47 +83,67 @@ class Reserve extends Component{
   }
 
   render(){
+    const style={
+    //   backgroundColor: '#4D4D4D',
+      padding: '10px',
+    //   fontWeight: 'bold',
+      width: '90vw',
+      maxWidth: '750px',
+      marginTop: '5px',
+      marginBottom: '5px',
+    }
+    const fieldset={
+      border: 'none',
+      borderLeft: '2px solid #E66E1C',
+      borderRight: '2px solid #E66E1C',
 
+    }
+    const legendStyle={
+      textDecoration: 'overline underline',
+      border: '10px #F4BE41',
+      borderWidth: '10px',
+      backgroundColor: 'white',
+      textAlign: 'center',
+      fontSize: '40px',
+      fontWeight: '200',
+      color: '#3C5BBE',
+      marginBottom: '25px'
+    }
     const labelStyle={
     //   border: "2px solid #383838",
     //   borderTop: "2px solid red",
       backgroundColor: "white",
-      padding: "10px 0px",
-      display: "flex",
-      alignItems: "center",
-      color: "#777",
-      textTransform:"uppercase"
+
+      padding: "10px 30px",
+      display: "flex"
     }
     const inputStyle={
-        height: "35px",
+        height: "18px",
         flexGrow: "1",
         marginLeft: "6px",
         border: "1px solid #ccc",
-        fontSize: "15px",
+        boxShadow: "inset 0 1px 3px #ddd",
+        borderRadius: "4px",
+        fontSize: "12px"
       }
       const inputButtonStyle={
           marginTop: '25px',
-          color: "#777",
-          fontWeight: "900",
-          backgroundColor: "#3675B6",
-          padding: '5px 15px',
-          color: "white",
-          textTransform: "uppercase"
+          borderRadius: "4px"
       }
     return(
-      <div className="reserve">
-        <fieldset>
-          <h1>Reserve Your Room</h1>
-            <div style={labelStyle}>Room Id:
-              <input id="tokenId" type="text" selected="true"style={inputStyle} onChange={this.handleTextChange} value={this.state.tokenId} />
+      <div style={style} className="reserve">
+        <fieldset style={fieldset}>
+          <legend style={legendStyle}>Reserve Your Room</legend>
+            <div style={labelStyle}>Room Id: 
+              <input id="tokenId" type="text" style={inputStyle} onChange={this.handleTextChange} value={this.state.tokenId} />
             </div>
-            <div style={labelStyle}> Check-in date:
+            <div style={labelStyle}> Check-in date: 
               <input id="start" type="text" style={inputStyle} onChange={this.handleTextChange} value={this.state.start} />
             </div>
-            <div style={labelStyle}> Check-out date:
+            <div style={labelStyle}> Check-out date: 
               <input id="stop" type="text" style={inputStyle} onChange={this.handleTextChange} value={this.state.stop} />
             </div>
-            <div style={labelStyle}> Access Code:
+            <div style={labelStyle}> Access Code: 
               <input id="accessCode" type="text" style={inputStyle} onChange={this.handleTextChange} value={this.state.accessCode} />
             </div>
             {/* <hr /> */}
