+++ src/containers/Steps.jsx
@@ -8,15 +8,14 @@ import colors from '../styles/colors';
 
 export default class Steps extends Component {
     render() {
-
         return (
             <div className="steps">
-                <Section
+                <Section 
                     pad='none'
                     align='stretch'
                     alignContent='center'
-                    // appCentered='true'
-                >
+                    appCentered='true'
+                > 
 
                 {/* 5 */}
                 <div className='heading'>Steps</div>
@@ -25,7 +24,7 @@ export default class Steps extends Component {
                 </p>
 
                 {/* 5.1 */}
-                <Heading align='center'
+                <Heading align='center' 
  	 	 	 	 	 	 margin='medium'
                         margin='medium'
                         tag='h2'>
@@ -37,7 +36,7 @@ export default class Steps extends Component {
                 </p>
 
                 {/* 5.1.1 */}
-                <Heading
+                <Heading 
                     align='center'
                     margin='medium'
                     tag='h4'>
@@ -48,13 +47,16 @@ export default class Steps extends Component {
                     token (as introduced in ERC-721) to provide temporary access for non-owners.
                     Applied to a hotel, our model does the following:
                 </p>
-                <CustomList
+                <CustomList 
                     bgColor={colors.gray}
-                >
+                > 
                     <CustomListItem
                         //height="auto"
                         // animation="slide-left"
                         // color="$black"
+                        // bgColor="orange"
+                        border="1px solid blue"
+                        borderRad
                         classNames="steps__card-font"
                     >
                         <div>
@@ -67,7 +69,8 @@ export default class Steps extends Component {
                         // animation="slide-left"
                         // color="$black"
                         // bgColor="orange"
-
+                        border="1px solid blue"
+                        borderRad
                         classNames="steps__card-font"
                     >
                         <div>
@@ -80,7 +83,8 @@ export default class Steps extends Component {
                         animation="slide-left"
                         // color="$black"
                         // bgColor="orange"
-
+                        border="1px solid blue"
+                        borderRad
                         classNames="steps__card-font"
                     >
                         <div>
@@ -95,13 +99,13 @@ export default class Steps extends Component {
                     </p>
 
                 {/* 5.1.2 */}
-                <Heading
+                <Heading 
                     align='center'
                     margin='medium'
                     tag='h2'>
                     Proof of Concept
                 </Heading>
-                <Heading
+                <Heading 
                     align='center'
                     margin='medium'
                     tag='h3'>
@@ -111,8 +115,8 @@ export default class Steps extends Component {
                     The minimum viable product will extend the Proof of Concept to incorporate
                     full payment functionality and user interfaces.
                 </p>
-
-                <Heading align='center'
+                
+                <Heading align='center' 
                     margin='medium'
                     tag='h3'>
                     1. Multisignature Wallets
@@ -121,8 +125,8 @@ export default class Steps extends Component {
                     To ensure secure payments, we will use industry standard code for our
                     multisignature wallet.
                 </p>
-
-                <Heading align='center'
+                   
+                <Heading align='center' 
                     margin='medium'
                     tag='h3'>
                     2. Traveler Interface
@@ -141,7 +145,8 @@ export default class Steps extends Component {
                         // animation="slide-left"
                         // color="$black"
                         // bgColor="orange"
-
+                        border="1px solid blue"
+                        borderRad
                         classNames="steps__card-font"
                     >
                         <div>
@@ -153,7 +158,8 @@ export default class Steps extends Component {
                         // animation="slide-left"
                         // color="$black"
                         // bgColor="orange"
-
+                        border="1px solid blue"
+                        borderRad
                         classNames="steps__card-font"
                     >
                         <div>
@@ -165,7 +171,8 @@ export default class Steps extends Component {
                         // animation="slide-left"
                         // color="$black"
                         // bgColor="orange"
-
+                        border="1px solid blue"
+                        borderRad
                         classNames="steps__card-font"
                     >
                         <div>
@@ -177,7 +184,8 @@ export default class Steps extends Component {
                         animation="slide-left"
                         // color="$black"
                         // bgColor="orange"
-
+                        border="1px solid blue"
+                        borderRad
                         classNames="steps__card-font"
                     >
                         <div>
@@ -188,7 +196,9 @@ export default class Steps extends Component {
                         //height="auto"
                         animation="slide-left"
                         // color="$black"
-
+                        // bgColor="orange"
+                        border="1px solid blue"
+                        borderRad
                         classNames="steps__card-font"
                     >
                         <div>
@@ -197,7 +207,7 @@ export default class Steps extends Component {
                     </CustomListItem>
                 </CustomList>
 
-                <Heading align='center'
+                <Heading align='center' 
                     margin='medium'
                     tag='h3'>
                     3. Hotel Interface
@@ -219,7 +229,8 @@ export default class Steps extends Component {
                         // animation="slide-left"
                         // color="$black"
                         // bgColor="orange"
-
+                        border="1px solid blue"
+                        borderRad
                         classNames="steps__card-font"
                     >
                         <div>
@@ -230,7 +241,9 @@ export default class Steps extends Component {
                         //height="auto"
                         // animation="slide-left"
                         // color="$black"
-
+                        // bgColor="orange"
+                        border="1px solid blue"
+                        borderRad
                         classNames="steps__card-font"
                     >
                         <div>
@@ -240,7 +253,10 @@ export default class Steps extends Component {
                     <CustomListItem
                         //height="auto"
                         animation="slide-left"
-
+                        // color="$black"
+                        // bgColor="orange"
+                        border="1px solid blue"
+                        borderRad
                         classNames="steps__card-font"
                     >
                         <div>
@@ -252,7 +268,7 @@ export default class Steps extends Component {
                         animation="slide-left"
                         // color="$black"
                         // bgColor="orange"
-                        border="1px solid $bl-blue"
+                        border="1px solid blue"
                         borderRad
                         classNames="steps__card-font"
                     >
@@ -261,11 +277,11 @@ export default class Steps extends Component {
                         </div>
                     </CustomListItem>
                 </CustomList>
-
+                
 
 
                 {/* 5.2 */}
-                <Heading align='center'
+                <Heading align='center' 
  	 	 	 	 	 	 margin='medium'
                     tag='h2'>
                     Early Adopters
@@ -278,7 +294,7 @@ export default class Steps extends Component {
                 </p>
 
                 {/* 5.2.1 */}
-                <Heading align='center'
+                <Heading align='center' 
  	 	 	 	 	 	 margin='medium'
                     tag='h3'>
                     Free interface for travelers.
@@ -289,7 +305,7 @@ export default class Steps extends Component {
                 </p>
 
                 {/* 5.2.2 */}
-                <Heading align='center'
+                <Heading align='center' 
  	 	 	 	 	 	 margin='medium'
                     tag='h3'>
                     Free interface for individual hosts.
@@ -300,7 +316,7 @@ export default class Steps extends Component {
                 </p>
 
                 {/* 5.2.3 */}
-                <Heading align='center'
+                <Heading align='center' 
  	 	 	 	 	 	 margin='medium'
                     tag='h3'>
                     Paid interface for hotels.
@@ -312,7 +328,7 @@ export default class Steps extends Component {
                 </p>
 
                 {/* 5.3 */}
-                <Heading align='center'
+                <Heading align='center' 
  	 	 	 	 	 	 margin='medium'
                     tag='h2'>
                     More Adopters
