+++ src/containers/what/What.jsx
@@ -2,8 +2,19 @@ import React, { Component } from 'react';
 import Paragraph from 'grommet/components/Paragraph';
 import Section from 'grommet/components/Section';
 import Heading from 'grommet/components/Heading';
+import Image from 'grommet/components/Image';
 import CustomList from '../CustomList';
 import CustomListItem from '../CustomListItem';
+import Search from '../../icons/search.png';
+import Compare from '../../icons/git-compare.png';
+import Reception from '../../icons/hotel_reception-512-bw.png';
+import Housekeeping from '../../icons/housekeeping-round-bw.svg';
+import Revenue from '../../icons/revenue-management.png';
+import Reserve from '../../icons/reserve-book.png';
+import ShoppingCart from '../../icons/shopping-cart.png';
+import Sale from '../../icons/point-of-sale.png';
+import Report from '../../icons/report.png';
+import Channel from '../../icons/channel-management.png';
 import colors from '../../styles/colors';
 
 class What extends Component {
@@ -11,15 +22,15 @@ class What extends Component {
   render() {
     return (
       <div className='what'>
-        <Section
+        <Section 
           pad='none'
           align='stretch'
           alignContent='center'
           // appCentered='true'
-        >
+        > 
 
           {/* 1 */}
-          <div className='heading'>What Do we Do?</div>
+          <div className='heading'>What</div>
           <p className='paragraph-white'>
             The vision for BookLocal is to connect travelers directly to their hotel of choice
             by creating the first two way property management system accessible as both
@@ -38,76 +49,83 @@ class What extends Component {
           <p className='paragraph-white'>
             To the traveler, BookLocal will provide the following features:
           </p>
-
-          <CustomList
+          
+          <CustomList 
             bgColor={colors.gray}
-
+            
           >
-            <CustomListItem
+            <CustomListItem 
               // height="auto"
               // animation="slide-left"
               color="$black"
               // bgColor="orange"
-              //border={`1px solid ${colors.blue1}`}
-              //borderRad="5px"
-              classNames="what__card-font">
-              {/* <SearchIcon/>   <Image
-                  size='thumb'
-                  src={Search}
-                /> */}
-
-              <h3><i class="fas fa-search"></i> Search</h3>
+              border={`1px solid ${colors.blue1}`}
+              borderRad="5px"
+              classNames="what__card-font"
+            >
+              {/* <SearchIcon/> */}
+              <Image
+                size='thumb'
+                src={Search}
+              />
+              Search
+              <div></div>
             </CustomListItem>
             <CustomListItem
               //height="auto"
               // animation="slide-left"
               // color="$black"
               // bgColor="orange"
-            //  border="1px solid blue"
-            //  borderRad
-              classNames="what__card-font">
-              {/* <CompareIcon/>   <Image
-                  size='thumb'
-                  src={Compare}
-                /> */}
-
-              <h3><i class="fas fa-exchange-alt"></i> Compare</h3>
+              border="1px solid blue"
+              borderRad
+              classNames="what__card-font"
+            >
+              {/* <CompareIcon/> */}
+              <Image
+                size='thumb'
+                src={Compare}
+              />
+              Compare
+              <div></div>
             </CustomListItem>
             <CustomListItem
               //height="auto"
               animation="slide-left"
               // color="$black"
               // bgColor="orange"
-            //  border="1px solid blue"
-              //borderRad
+              border="1px solid blue"
+              borderRad
               classNames="what__card-font"
             >
-              {/* <CheckboxSelectedIcon/>   <Image
-                  size='thumb'
-                  src={Reserve}
-                />*/}
-
-              <h3><i class="far fa-address-book"></i> Book</h3>
+              {/* <CheckboxSelectedIcon/> */}
+              <Image
+                size='thumb'
+                src={Reserve}
+              />
+              Book
+              <div></div>
             </CustomListItem>
             <CustomListItem
               //height="xsmall"
               animation="slide-left"
               // color="$black"
               // bgColor="orange"
-            //  border="1px solid blue"
-              //borderRad
-              classNames="what__card-font">
-              {/* <MoneyIcon/>     <Image
-                    size='thumb'
-                    src={ShoppingCart}
-                  /> */}
-
-              <h3><i class="fas fa-shopping-cart"></i> Purchase</h3>
+              border="1px solid blue"
+              borderRad
+              classNames="what__card-font"
+            >
+              {/* <MoneyIcon/> */}
+              <Image
+                size='thumb'
+                src={ShoppingCart}
+              />
+              Purchase
+              <div></div>
             </CustomListItem>
           </CustomList>
-
-            <h2>To the hotel, BookLocal will provide modules for:</h2>
-
+          <p className='paragraph-white'>
+            To the hotel, BookLocal will provide modules for:
+          </p>
           {/* add icons */}
           <CustomList
             bgColor={colors.gray}
@@ -117,89 +135,96 @@ class What extends Component {
               // animation="slide-left"
               // color="$black"
               // bgColor="orange"
-              //border="1px solid blue"
-              //borderRad
-              classNames="what__card-font">
-              {/* <Image
+              border="1px solid blue"
+              borderRad
+              classNames="what__card-font"
+            >
+              <Image
                 size='thumb'
                 src={Reception}
-              />*/}
-              <h3><i class="fas fa-users"></i> Guest Management</h3>
+              />
+              Guest Management
+              <div></div>
             </CustomListItem>
             <CustomListItem
               // height="xsmall"
               // animation="slide-left"
               // color="$black"
               // bgColor="orange"
-              //border="1px solid blue"
-            //  borderRad
+              border="1px solid blue"
+              borderRad
               classNames="what__card-font"
             >
-          {    /*<Image
+              <Image
                 size='thumb'
                 src={Housekeeping}
-              />*/}
-              <h3><i class="fas fa-home"></i> Housekeeping and Maintenance</h3>
+              />
+              Housekeeping and Maintenance
+              <div></div>
             </CustomListItem>
             <CustomListItem
               // height="xsmall"
               // animation="slide-left"
               // color="$black"
               // bgColor="orange"
-              // border="1px solid blue"
-              // borderRad
+              border="1px solid blue"
+              borderRad
               classNames="what__card-font"
             >
-              {/*<Image
+              <Image
                 size='thumb'
                 src={Revenue}
-              />*/}
-              <h3><i class="fas fa-money-bill-alt"></i> Revenue Management</h3>
+              />              
+              Revenue Management
+              <div></div>
             </CustomListItem>
             <CustomListItem
               // height="xsmall"
               animation="slide-left"
               // color="$black"
               // bgColor="orange"
-              //border="1px solid blue"
-              // borderRad
+              border="1px solid blue"
+              borderRad
               classNames="what__card-font"
             >
-              {/*<Image
+              <Image
                 size='thumb'
                 src={Sale}
-              />*/}
-               <h3><i class="far fa-credit-card"></i> Payment Processing and Point of Sale</h3>
+              />
+              Payment Processing and Point of Sale
+              <div></div>
             </CustomListItem>
             <CustomListItem
                 // height="xsmall"
                 animation="slide-left"
                 // color="$black"
                 // bgColor="orange"
-                //border="1px solid blue"
-                //borderRad
+                border="1px solid blue"
+                borderRad
                 classNames="what__card-font"
               >
-              {/*<Image
+              <Image
                 size='thumb'
                 src={Report}
-              />*/}
-              <h3><i class="fas fa-chart-line"></i> Report Generation</h3>
+              />
+              Report Generation
+              <div></div>
             </CustomListItem>
             <CustomListItem
-
+              // height="xsmall"
               animation="slide-left"
               // color="$black"
               // bgColor="orange"
-              //border="1px solid blue"
-              //borderRad
+              border="1px solid blue"
+              borderRad
               classNames="what__card-font"
             >
-            {  /*<Image
+              <Image
                 size='thumb'
                 src={Channel}
-              />*/}
-              <h3><i class="fas fa-arrows-alt"></i> Channel Management</h3>
+              />
+              Channel Management
+              <div></div>
             </CustomListItem>
           </CustomList>
         </Section>
