+++ src/containers/why/Problems.jsx
@@ -9,13 +9,15 @@ class Problems extends Component {
     return (
       <div className="problems">
         {/* 2.1 */}
-        <Heading
-
-          tag='h2'
+        <Heading 
+          align='center' 
+ 	 	 	 	 	margin='large'
+          tag='h1'
+          style={{color: colors.logoBlue}}
         >
           Problems
         </Heading>
-        <p className='paragraph-white'>
+        <p className='paragraph-gray'>
           The hotel booking industry is fragmented with no fewer than five different
           groups working for commission between travelers and hotels. Problems include:
         </p>
